refactor(desktop): split composer model selector logic
