+++ apps/desktop/src/lib/domain/chat/models/model-selector-types.ts
@@ -1,0 +1,58 @@
+import type { AgentSummary } from "@anyharness/sdk";
+import type { PendingSessionConfigChangeStatus } from "@proliferate/product-domain/sessions/pending-config";
+import type { ChatModelVisibilityOverridesByAgentKind } from "@/lib/domain/preferences/user/session-defaults";
+
+export interface ModelSelectorSelection {
+  kind: string;
+  modelId: string;
+}
+
+export interface ChatLaunchPreferences {
+  defaultChatAgentKind: string;
+  defaultChatModelIdByAgentKind: Record<string, string>;
+  chatModelVisibilityOverridesByAgentKind?: ChatModelVisibilityOverridesByAgentKind;
+}
+
+export type ModelSelectionActionKind =
+  | "select"
+  | "update_current_chat"
+  | "open_new_chat";
+
+export interface ModelSelectorItem {
+  kind: string;
+  modelId: string;
+  displayName: string;
+  actionKind: ModelSelectionActionKind;
+  isSelected: boolean;
+}
+
+export interface ModelSelectorGroup {
+  kind: string;
+  providerDisplayName: string;
+  models: ModelSelectorItem[];
+}
+
+export interface ActiveModelSelectorControl {
+  kind: string;
+  values: ReadonlyArray<{
+    value: string;
+    label: string;
+    description?: string | null;
+  }>;
+}
+
+export interface ModelSelectorCurrentModel {
+  kind: string;
+  displayName: string;
+  pendingState: PendingSessionConfigChangeStatus | null;
+}
+
+export interface ModelSelectorProps {
+  connectionState: string;
+  currentModel: ModelSelectorCurrentModel | null;
+  groups: ModelSelectorGroup[];
+  hasAgents: boolean;
+  isLoading: boolean;
+  notReadyAgents: AgentSummary[];
+  onSelect: (selection: ModelSelectorSelection) => void;
+}
