+++ scripts/max_lines_allowlist.txt
@@ -22,7 +22,6 @@ anyharness/sdk-react/src/hooks/sessions.ts 653 existing SDK React oversized file
 anyharness/sdk/src/reducer/__tests__/transcript.test.ts 1880 existing SDK oversized test file
 anyharness/sdk/src/reducer/transcript.ts 1575 existing SDK oversized file
 apps/desktop/src-tauri/src/commands/ssh_tunnel.rs 729 existing desktop native oversized file
-apps/desktop/src/components/workspace/chat/input/ComposerModelConfigSelector.tsx 735 existing desktop component oversized file
 apps/desktop/src/components/workspace/shell/right-panel/RightPanel.test.tsx 732 existing desktop component oversized test file
 apps/desktop/src/lib/domain/workspaces/cloud/logical-workspaces.test.ts 759 existing desktop oversized test file
 apps/desktop/src/lib/workflows/mcp/connector-persistence.test.ts 704 existing desktop oversized test file
